Add files via upload
Use this script to read and process data for any area of interest using its shape file (.shp). This script is particularly useful for hydrologist working on river basins. This file is self-explanatory in nature. For more clarity of syntax, use of online chatbots like ChatGPT is recommended for customisation i.e to suit structure of your .nc file data.